requirements.txt
requests~=2.31.0
SQLAlchemy~=2.0.25
pydantic~=2.5.3
flask~=3.0.0
psycopg2-binary==2.9.9